requirements.txt
fastapi
uvicorn
pandas
streamlit
google-generativeai
pandas
fpdf
python-dotenv
imageio-ffmpeg
setuptools
